Refactor: New options menu design, New options script
